requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
Django==4.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filestack-python==3.5.0
idna==3.4
pycparser==2.21
pycryptodome==3.17
PyNaCl==1.5.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
trafaret==2.0.2
tzdata==2023.3
urllib3==1.26.15
